+++ model-test/src/main/java/org/jboss/as/model/test/MavenUtil.java
@@ -448,13 +448,13 @@ class MavenUtil {
         @Override
         public void transferProgressed(TransferEvent event) {
             TransferResource resource = event.getResource();
-            downloads.put(resource, Long.valueOf(event.getTransferredBytes()));
+            downloads.put(resource, event.getTransferredBytes());
 
             StringBuilder buffer = new StringBuilder(64);
 
             for (Map.Entry<TransferResource, Long> entry : downloads.entrySet()) {
                 long total = entry.getKey().getContentLength();
-                long complete = entry.getValue().longValue();
+                long complete = entry.getValue();
 
                 buffer.append(getStatus(complete, total)).append("  ");
             }
